pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Konstantin Pavlov <k.pavlov.reg@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"
pyyaml = "^6.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.1.0a6"]
build-backend = "poetry.core.masonry.api"
